fix: Update commit calculations to be accurate
When forking a repo, all the commits in the forked repository gets
counted when grabbing data from:

/repos/{owner}/{repo}/stats/commit_activity .

As a result, the old way of calculations led to an unusually high
commit count. Replace it by grabbing data from contributors instead
and check the age of the contribution week.

https://docs.github.com/en/rest/reference/repos#get-all-contributor-commit-activity

Note: the contributors endpoint counts commits per week. It might
take a week of commits off the count.