fix: Update repo data fetching for correctness
For 'List repositories for the authenticated user', the default
results per page is 30. If you have more than 30 repos, they will
not be included in the return of grabDataFromAllRepositories().

Instead, use an options object and pass it as an argument to
octokit.rest.repos.listForAuthenticatedUser(). Note: this solution
only includes up to 100 repos, which works for me. Maybe use
octokit.paginate() if you have more than 100.

- Replace octokit.request() with octokit.request.repos.get/list
  for default options.

- Improve readability by simplifying operations and using
  functions.